remove features
