+++ packages/kit/src/views/Onboarding/pages/ConnectHardwareWallet/ConnectYourDevice.tsx
@@ -647,7 +647,6 @@ function ConnectByUSBOrBLE({
       ) {
         await showFirmwareVerifyDialog({
           device,
-          features,
           onContinue: async ({ checked }) => {
             if (deviceMode === EOneKeyDeviceMode.notInitialized) {
               handleNotActivatedDevicePress({ deviceType });
